Move trait bounds to HRTBs on the Backend trait definition.
This actually does not solve the issue of proliferating trait bounds because of a
known issue in Rust's trait bound resolution system. It's being addressed and the
only solution for now appears to be to repeat the bounds absolutely everywhere
they are relevant or perhaps to implement a custom trait that implements these bounds
for the specific associated type rather than use a where clause.